Renamed main.py to be transcript.py and created a new main.py as a demo for the transcript class
