+++ src/genbank_seqkit/main.py
@@ -1,259 +1,68 @@
-import logging
+#!/usr/bin/env python3
+"""
+Main entry point for the Genbank Seqkit project.
 
-from genbank_seqkit.utils.entrez_efetch import fetch_transcript_record
+Demonstrates usage of the Transcript class, fetching GenBank records,
+and generating FASTA/GenBank-style outputs.
+"""
+
+import sys
 from genbank_seqkit.logger import logger
-from genbank_seqkit.utils import _force_list
 from genbank_seqkit.errors import (
     TranscriptIdError,
     GenbankFetchError,
     GenbankParseError,
     GenbankError
 )
+from genbank_seqkit.main import Transcript  # Your Transcript class
 
-class Transcript:
+def main(transcript_id: str, verbose: bool = False):
     """
-    Represents a GenBank transcript record. Automatically fetches data from NCBI and populates key attributes.
+    Fetch a transcript from NCBI and print summary, FASTA, and GenBank formats.
+
+    Parameters
+    ----------
+    transcript_id : str
+        RefSeq transcript ID (e.g., "NM_000093.5").
+    verbose : bool, optional
+        If True, prints debug logging information.
     """
-    def __init__(self, transcript_id: str, verbose=False):
-        # Core attributes
-        self.transcript_id = transcript_id
-        self.gene_symbol = None
-        self.hgnc_id = None
-        self.dna_sequence = None
-        self.rna_sequence = None
-        self.protein_sequence = None
-        self.protein_id = None
-        self.length = None
-        self.mol_type = None
-        self.gene_name = None
+    try:
+        # Create Transcript object (fetches and populates automatically)
+        transcript = Transcript(transcript_id, verbose=verbose)
 
-        # Fetch and populate attributes automatically
-        self._fetch_and_populate(verbose=verbose)
+        # Print user-friendly summary
+        print(transcript)
 
-    def _fetch_and_populate(self, verbose=False):
-        """
-        Internal method to fetch the GenBank record using entrez_efetch and populate key attributes.
+        # Generate FASTA output
+        fasta_str = transcript.as_fasta(seq_type="DNA")
+        print("\nFASTA format:\n", fasta_str)
 
-        Parameters
-        ------------
-        verbose: bool, optional
-            If True, logs detailed messages about each step of the population process.
+        # Generate simple GenBank output
+        genbank_str = transcript.as_genbank(seq_type="DNA")
+        print("\nGenBank format:\n", genbank_str)
 
-        Raises
-        ------------
-        TranscriptIdError
-            If an invalid transcript ID is provided.
-        GenbankFetchError
-            If fetching the record from NCBI fails.
-        GenbankParseError
-            If parsing the XML record fails.
-        GenbankError
-            For any unexpected error encountered during population.
-        """
+    except TranscriptIdError as e:
+        logger.error(f"Invalid transcript ID: {e}")
+        print(f"Error: {e}")
 
-        # Logger depends on if verbose is True or not
-        temp_logger = logger if verbose else logging.getLogger("null_logger")
+    except (GenbankFetchError, GenbankParseError) as e:
+        logger.error(f"Problem retrieving/parsing transcript {transcript_id}: {e}")
+        print(f"Error: {e}")
 
-        try:
-            record = fetch_transcript_record(self.transcript_id)
-            temp_logger.debug(f"Fetched transcript record {self.transcript_id}")
+    except GenbankError as e:
+        logger.error(f"Unexpected GenBank error for {transcript_id}: {e}")
+        print(f"Error: {e}")
 
-            # Top-level fields:
-            self.transcript_id = record.get('GBSeq_accession-version', self.transcript_id)
-            self.dna_sequence = record['GBSeq_sequence'].upper()
-            self.rna_sequence = self.dna_sequence.replace('T', 'U')
-            self.length = record.get("GBSeq_length")
-            self.mol_type = record.get("GBSeq_moltype")
-            self.gene_name = record.get("GBSeq_definition")
-            temp_logger.debug(f"Top-level attributes loaded: length={self.length}, mol_type={self.mol_type}, "
-                              f"gene name={self.gene_name}")
-
-            # Parse feature table
-            features = _force_list(record.get("GBSeq_feature-table", {}).get("GBFeature"), verbose=False)
-            if not features:
-                temp_logger.warning(f"No features found for transcript {self.transcript_id}")
-
-            # Find Qualifiers within feature table and define as 'quals' and Feature key as 'key'
-            for feature in features:
-                key = feature.get("GBFeature_key")
-                quals = _force_list(feature.get("GBFeature_quals", {}).get("GBQualifier"), verbose=False)
-
-                # Gene symbol and HGNC ID
-                if key == "gene":
-                    for q in quals:
-                        if q.get("GBQualifier_name") == "gene":
-                            self.gene_symbol = q.get("GBQualifier_value")
-                            temp_logger.debug(f"Gene symbol found: {self.gene_symbol}")
-                        elif q.get("GBQualifier_name") == "db_xref":
-                            val = q.get("GBQualifier_value", "")
-                            if val.startswith("HGNC:"):
-                                self.hgnc_id = val.replace("HGNC:", "")
-                                temp_logger.debug(f"HGNC ID found: {self.hgnc_id}")
-
-                # CDS and protein info
-                elif key == "CDS":
-                    for q in quals:
-                        if q.get("GBQualifier_name") == "translation":
-                            self.protein_sequence = q.get("GBQualifier_value")
-                            temp_logger.debug(f"Protein sequence found, length={len(self.protein_sequence)}")
-                        elif q.get("GBQualifier_name") == "protein_id":
-                            self.protein_id = q.get("GBQualifier_value")
-                            temp_logger.debug(f"Protein ID found: {self.protein_id}")
-
-            # Log any missing attributes
-            if not self.gene_symbol:
-                temp_logger.warning(f"No gene symbol found for transcript {self.transcript_id}")
-            if not self.hgnc_id:
-                temp_logger.warning(f"No hgnc ID found for transcript {self.transcript_id}")
-            if not self.protein_sequence:
-                temp_logger.warning(f"No protein sequence found for transcript {self.transcript_id}")
-            if not self.protein_id:
-                temp_logger.warning(f"Protein ID missing for {self.transcript_id}")
-
-        except (TranscriptIdError, GenbankFetchError, GenbankParseError) as e:
-            logger.error(f"Error processing transcript {self.transcript_id}: {e}")
-            raise
-
-        except Exception as e:
-            logger.error(f"Failed to fetch transcript {self.transcript_id}: {e}")
-            raise GenbankError(f"Failed to populate transcript {self.transcript_id}") from e
-            raise
-
-    def refresh(self, verbose=False):
-        """ Public method for re-fetching and updating the transcript. """
-        self._fetch_and_populate(verbose=verbose)
-
-    def as_fasta(self, sequence=None, seq_type="DNA"):
-        """
-        Return FASTA formatted sequence string for the given sequence.
-        Defaults to the DNA transcript sequence if none is provided.
-        Logs warnings if the sequence is missing and errors for unknown types.
-
-        Parameters
-        ------------
-        sequence : str, optional
-            The sequence to format. Defaults to self.dna_sequence.
-        seq_type: str, optional
-            The type of sequence to return, can be "DNA", "RNA" or "protein". Defaults to "DNA".
-
-        Returns
-        ------------
-        str
-            FASTA formatted string.
-
-        Raises
-        ------------
-        ValueError
-            If unknown seq_type is input.
-
-        """
-        if sequence is None:
-            if seq_type.upper() == "DNA":
-                sequence = self.dna_sequence
-            elif seq_type.upper() == "RNA":
-                sequence = self.rna_sequence
-            elif seq_type.upper() == "protein":
-                sequence = self.protein_sequence
-            else:
-                logger.error(f"Unknown sequence type requested: {seq_type}")
-                raise ValueError(f"Unknown seq_type: {seq_type}")
-
-        if sequence is None or sequence == "":
-            logger.warning(f"{seq_type} sequence not available for {self.transcript_id}")
-            sequence = ""
-
-        header = f">{self.transcript_id} | {seq_type}"
-        fasta_str = f"{header}\n{sequence}"
-
-        logger.debug(f"Generated FASTA formatted sequence for {self.transcript_id}, seq_type: {seq_type}")
-        return fasta_str
-
-    def as_genbank(self, sequence=None, seq_type="DNA"):
-        """
-        Return a simple GenBank-like formatted string.
-        Defaults to the DNA transcript sequence if none is provided.
-
-        Parameters
-        ------------
-        sequence: str, optional
-            The sequence to format. Defaults to self.dna_sequence.
-        seq_type: str, optional
-            The type of sequence to return, can be "DNA", "RNA" or "protein". Defaults to "DNA".
-
-        Returns
-        ------------
-        str
-            Simple GenBank-like formatted string.
-
-        Raises
-        ------------
-        ValueError: If unknown seq_type is input.
-
-        """
-        if sequence is None:
-            if seq_type.upper() == "DNA":
-                sequence = self.dna_sequence
-            elif seq_type.upper() == "RNA":
-                sequence = self.rna_sequence
-            elif seq_type.upper() == "protein":
-                sequence = self.protein_sequence
-            else:
-                logger.error(f"Unknown sequence type requested: {seq_type}")
-                raise ValueError(f"Unknown seq_type: {seq_type}")
-
-        if sequence is None or sequence == "":
-            logger.warning(f"{seq_type} sequence not available for {self.transcript_id}")
-            sequence = ""
-
-        genbank_str = (
-            f"LOCUS        {self.transcript_id}\n"
-            f"DEFINITION   {seq_type} sequence\n"
-            f"ORIGIN\n{sequence}\n//"
-        )
-
-        logger.debug(f"Generated GenBank string for {self.transcript_id}, seq_type: {seq_type}")
-        return genbank_str
-
-    def __repr__(self):
-        """
-        Developer-friendly representation for the REPL or debugging.
-        Shows the class name and transcript ID.
-        """
-        return f"<Transcript {self.transcript_id}>"
-
-    def __str__(self):
-        """
-        User-friendly string representation.
-        Includes transcript ID, gene symbol (if available), and length.
-        """
-        gene = self.gene_symbol if self.gene_symbol else "Unknown"
-        length = self.length if self.length else "Unknown"
-        return f"Transcript: {self.transcript_id} | Gene: {gene} | Length: {length} bp"
-
-# Demo
-if __name__ == "__main__": # pragma: no cover
-
-    ## Use this code block to create a .json file for development.
-    # import pprint
-    # import json
-    #
-    # record = fetch_transcript_record("NM_000093.5")
-    # pprint.pprint(record)
-    #
-    # with open("record.json", "w") as f:
-    #     json.dump(record,f, indent = 2)
+    except Exception as e:
+        logger.error(f"Unexpected failure for {transcript_id}: {e}")
+        print(f"An unexpected error occurred: {e}")
 
 
-    t = Transcript("NM_000093.5", verbose=True)
-    # print(t)
-    # print(t.transcript_id)
-    # print(t.dna_sequence[:50])  # first 50 bases
-    # print(t.as_fasta(t.dna_sequence))
-    # print(t.length)
-    # print(t.mol_type)
-    # print(t.gene_symbol)
-    # print(t.protein_sequence)
-    # print(t.hgnc_id)
-    # print(t.as_fasta())
-    # print(t.as_genbank(seq_type="RNA"))
+if __name__ == "__main__":
+    if len(sys.argv) < 2:
+        print("Usage: python main.py <transcript_id>")
+        sys.exit(1)
 
+    transcript_id = sys.argv[1]
+    main(transcript_id, verbose=True)
